+++ astropy/convolution/tests/test_convolve_fft.py
@@ -1,8 +1,8 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 import numpy as np
 
-from ....tests.helper import pytest
-from ....config import ConfigurationItem
+from ...tests.helper import pytest
+from ...config import ConfigurationItem
 
 from ..convolve import convolve_fft