+++ astropy/nddata/convolution/__init__.py
@@ -1,3 +1,0 @@
-from .core import *
-from .kernels import *
-from .convolve import convolve, convolve_fft